+++ FakeServer/Common/Config.cs
@@ -7,6 +7,5 @@
         public const string GraphQLRoute = "graphql";
         public const string TokenRoute = "token";
         public const string TokenLogoutRoute = "logout";
-        public const string IdField = "id";
     }
 }