+++ FakeServer/Common/DatastoreSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace FakeServer.Common
+{
+    public class DataStoreSettings
+    {
+        public string IdField { get; set; } = "id";
+
+        public bool EagerDataReload { get; set; }
+    }
+}